demo: corrección de referencias, y ajustes menos en variables
