--- src/edu/prof/carlos/castillo/reto2/boletos/Main.java
@@ -1,8 +1,0 @@
-package edu.prof.carlos.castillo.reto2.boletos;
-import edu.prof.carlos.castillo.reto2.boletos.ui.CLI;
-
-public class Main {
-    public static void main(String[] args) {
-        CLI.showMenu();
-    }
-}
