+++ src/edu/prof/carlos/castillo/reto2/piramide/boletos/Main.java
@@ -1,0 +1,9 @@
+package edu.prof.carlos.castillo.reto2.piramide.boletos;
+
+import edu.prof.carlos.castillo.reto2.piramide.boletos.ui.CLI;
+
+public class Main {
+    public static void main(String[] args) {
+        CLI.showMenu();
+    }
+}
